Working on new puzzle type, and created variants for existing ones.
Adding "+" at the end of the puzzle type makes it so blocks wont insert until they are all in correct spot, causing you to have to deal with pieces INSIDE the solution blocking movement. Blocks+ and Control+ both work, Menu+ acts the same as Menu, and Items+ probably would break.